refactor: Import Matrix and PSDMatrix from torchjd.linalg throughout the codebase
