+++ src/torchjd/aggregation/_config.py
@@ -5,7 +5,7 @@
 import torch
 from torch import Tensor
 
-from torchjd._linalg import Matrix
+from torchjd.linalg import Matrix
 
 from ._aggregator_bases import Aggregator
 from ._sum import SumWeighting
