+++ src/torchjd/aggregation/_mgda.py
@@ -1,7 +1,7 @@
 import torch
 from torch import Tensor
 
-from torchjd._linalg import PSDMatrix
+from torchjd.linalg import PSDMatrix
 
 from ._aggregator_bases import GramianWeightedAggregator
 from ._weighting_bases import _GramianWeighting
