+++ src/torchjd/autogram/_gramian_computer.py
@@ -4,8 +4,9 @@ from typing import cast
 from torch import Tensor
 from torch.utils._pytree import PyTree
 
-from torchjd._linalg import Matrix, PSDMatrix, compute_gramian
+from torchjd._linalg import compute_gramian
 from torchjd.autogram._jacobian_computer import JacobianComputer
+from torchjd.linalg import Matrix, PSDMatrix
 
 
 class GramianComputer(ABC):
